LPS-70170 Move (no logic changes)
