+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/internal/JspCDefaultsPlugin.java
@@ -15,7 +15,6 @@
 package com.liferay.gradle.plugins.internal;
 
 import com.liferay.gradle.plugins.BasePortalToolDefaultsPlugin;
-import com.liferay.gradle.plugins.extensions.LiferayExtension;
 import com.liferay.gradle.plugins.internal.util.FileUtil;
 import com.liferay.gradle.plugins.internal.util.GradleUtil;
 import com.liferay.gradle.plugins.jasper.jspc.CompileJSPTask;
@@ -32,7 +31,6 @@ import org.gradle.api.Plugin;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
 import org.gradle.api.file.ConfigurableFileCollection;
-import org.gradle.api.file.CopySpec;
 import org.gradle.api.plugins.JavaPlugin;
 import org.gradle.api.tasks.bundling.Jar;
 
@@ -163,39 +161,6 @@ public class JspCDefaultsPlugin
 				}
 
 			});
-
-		Action<Task> taskAction = new Action<Task>() {
-
-			@Override
-			public void execute(Task task) {
-				final CompileJSPTask compileJSPTask = (CompileJSPTask)task;
-
-				Action<CopySpec> copySpecAction = new Action<CopySpec>() {
-
-					@Override
-					public void execute(CopySpec copySpec) {
-						copySpec.from(compileJSPTask.getDestinationDir());
-
-						LiferayExtension liferayExtension =
-							GradleUtil.getExtension(
-								project, LiferayExtension.class);
-
-						File destinationDir = new File(
-							liferayExtension.getLiferayHome() + "/work/" +
-								GradleUtil.getArchivesBaseName(project) + "-" +
-									project.getVersion());
-
-						copySpec.into(destinationDir);
-					}
-
-				};
-
-				project.copy(copySpecAction);
-			}
-
-		};
-
-		compileJSPTask.doLast(taskAction);
 	}
 
 	private File _getUnzippedJarDir(Project project) {
